Added GetOrders, added cleanUser()
