+++ src/test/java/CreateOrder.java
@@ -1,5 +1,6 @@
 import io.restassured.RestAssured;
 import org.json.JSONObject;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
@@ -61,7 +62,7 @@ public class CreateOrder {
     }
 
     @Test
-    @DisplayName("Unauth user create order")
+    @DisplayName("Unauthorized user create order")
     public void unauthUserCreateOrder() {
         String json = String.format("{\"ingredients\": [\"61c0c5a71d1f82001bdaaa6d\",\"61c0c5a71d1f82001bdaaa6f\"]}");
         given()
@@ -105,5 +106,29 @@ public class CreateOrder {
                 .then().assertThat()
                 .and()
                 .statusCode(500);
+    }
+
+    @After
+    public void cleanUser() {
+        if (login == null) {
+            return;
+        }
+        String json = String.format("{\"email\": \"%s\",\"password\":\"password\"}", login);
+        var response = given()
+                .header("Content-type", "application/json")
+                .body(json)
+                .when()
+                .post("/api/auth/login");
+        response.then().assertThat().statusCode(200);
+        JSONObject jsonObject = new JSONObject(response.getBody().asString());
+        String accessTokenToDelete = jsonObject.get("accessToken").toString();
+
+        given()
+                .header("Content-type", "application/json")
+                .header("Authorization", accessTokenToDelete)
+                .when()
+                .delete("/api/auth/user")
+                .then()
+                .statusCode(202);
     }
 }
