+++ src/test/java/CreateUserTest.java
@@ -1,3 +1,8 @@
+import io.restassured.path.json.JsonPath;
+import io.restassured.response.Response;
+import io.restassured.specification.RequestSpecification;
+import org.json.JSONObject;
+import org.junit.After;
 import org.junit.Before;
 import io.restassured.RestAssured;
 import org.junit.Test;
@@ -30,7 +35,7 @@ public class CreateUserTest {
     }
 
     @Test
-    @DisplayName ("Can not duplicate user")
+    @DisplayName("Can not duplicate user")
     public void cannotCreateDuplicateTest() {
         String json = getRandomUserJson();
         given()
@@ -46,10 +51,11 @@ public class CreateUserTest {
                 .when()
                 .post("/api/auth/register")
                 .then().statusCode(403);
+                login = null;
     }
 
     @Test
-    @DisplayName ("Can not create user without password")
+    @DisplayName("Can not create user without password")
     public void cannotCreateUserWithoutPasswordTest() {
         File json = new File("src/test/resources/newUserWithoutPassword.json");
         given()
@@ -58,12 +64,37 @@ public class CreateUserTest {
                 .when()
                 .post("/api/auth/register")
                 .then().statusCode(403);
+                login = null;
     }
 
-    private String getRandomUserJson(){
+    private String getRandomUserJson() {
         login = String.format("a%s@gmail.com", java.util.UUID.randomUUID());
         String json = String.format("{\"email\": \"%s\"," +
                 "  \"password\": \"password\", \"name\": \"Username\"}", login);
         return json;
+    }
+
+    @After
+    public void cleanUser() {
+        if (login == null) {
+            return;
+        }
+        String json = String.format("{\"email\": \"%s\",\"password\":\"password\"}", login);
+        var response = given()
+                .header("Content-type", "application/json")
+                .body(json)
+                .when()
+                .post("/api/auth/login");
+        response.then().assertThat().statusCode(200);
+        JSONObject jsonObject = new JSONObject(response.getBody().asString());
+        String accessTokenToDelete = jsonObject.get("accessToken").toString();
+
+        given()
+                .header("Content-type", "application/json")
+                .header("Authorization", accessTokenToDelete)
+                .when()
+                .delete("/api/auth/user")
+                .then()
+                .statusCode(202);
     }
 }
